code/requirements.txt
aniso8601==1.3.0
click==6.7
Flask==0.12.2
Flask-JWT==0.3.2
Flask-RESTful==0.3.6
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mysql-connector==2.1.4
PyJWT==1.4.2
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
Werkzeug==0.12.2
